Algorithm Update: -- Balanced approach to filling in remainder of crew members and gear following trip preferences. Makes it distribute more evenly -- Crew Members are now shuffled by position prior to being sorted allowing more positional distribution over all loads

+++ lib/Data/crewmember.dart
@@ -26,11 +26,12 @@ class CrewMember extends HiveObject{
     int totalWeight = flightWeight;
     if (personalTools != null) {
       for (var tool in personalTools!) {
-        totalWeight += tool.weight;
+        totalWeight += tool.weight.toInt(); // Ensure tool.weight is treated as int
       }
     }
-    return totalWeight;
+    return totalWeight; // Explicitly returning an int
   }
+
 
   CrewMember({required this.name, required this.flightWeight, required this.position, this.personalTools});
